refactor(qrcode): move utilities to shared @/lib/qrcode module
Extract QR code utilities to a dedicated shared module:
- limits.ts: Subscription-based QR code creation limits
- validation.ts: URL validation helpers
- index.ts: Barrel file for exports

Update imports in dashboard pages to use new module path
